LPS-169890 Inline the only usage
